requirements.txt
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
dlib==19.24.0
numpy==1.23.5
Pillow==9.3.0
tkinter
